start dron-n-chop rewrite as Leaflet subclasses
this is an example that takes an existing drop-n-chop "component"
and creates a DropZone Leaflet subclass with a FileReader
handle which delegates events for the map's drag-n-drop options.

I think it would be nice if the other drop-n-chop "components" were
also written as Leaflet extensions/subclasses.

For example:

- The FileLayerList could extend L.Control.Layers as this extensions
does >
https://github.com/stefanocudini/leaflet-panel-layers/blob/master/src/leaflet-panel-layers.js

- The GeoMenu could extend L.Control to create the dropdown toolbar

There would still be a drop-n-chop module that wraps the mapping needs
and also has refs to the above components as this commit shows with
L.DNC.DropZone.FileReader

regarding discussion in #8, we then don't need browserify at all.
regarding issues ( #22, #23 ), tests will be easier to write
when we come to an agreement on the structure