Rewrite drop-n-chop components with Leaflet API
This commit takes the existing 'components' and creates Leaflet
classes from them housed under the namespace L.DNC. The goal was
to see if this type of organization 'feels better' and to
learn more about the plugin atomosphere of Leaflet.
The drop-n-chop parts that follow this pattern include:

1. Dropzone

2. LayerList

3. Geoprocessing Menu

The remaining drop-n-chop application code is very tiny and IS
NOT a Leaflet subclass. It is just a simple JS class that
expects the Leaflet components to exist. Maybe this feels gross
for some people. This means there are two globals exposed:

window.L.DNC

window.DncApp

`browserify` is not needed anymore. The JS build just uses
`grunt-concat` to do package things in dep order.

@svmatthews and @alukach: I'd be interested to know if
this organization pattern feels right or wrong?